ft: add appointment date filters
