+++ .eslintrc.json
@@ -3,6 +3,7 @@
   "rules": {
     "@next/next/no-img-element": "off",
     "jsx-a11y/alt-text": "off",
-    "react/no-unescaped-entities": "off"
+    "react/no-unescaped-entities": "off",
+    " react-hooks/exhaustive-deps": "off"
   }
 }
